pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srp-gpt2-pytorch"
version = "0.1.0"
description = "A Single Responsibility Principle GPT-2 style Transformer implementation in PyTorch."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Generated by ChatGPT"}]
dependencies = [
  "torch>=2.1",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "numpy>=1.24",
]

[project.optional-dependencies]
tokenizers = ["tiktoken>=0.7"]
dev = ["pytest>=8.0", "ruff>=0.5", "mypy>=1.10"]

[project.scripts]
srp-gpt2 = "srp_gpt2.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests", "examples"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
